requirements.txt
Flask
Flask-PyMongo
Pillow
opencv-python
